Update and fix login/logout flows
