+++ admin-client/src/components/index.js
@@ -7,8 +7,8 @@ export { default as GridContainer } from './GridContainer.svelte';
 export { default as GridRow } from './GridRow.svelte';
 export { default as Hero } from './Hero.svelte';
 export { default as LabeledItem } from './LabeledItem.svelte';
-export { default as Logout } from './Logout.svelte';
 export { default as Nav } from './Nav.svelte';
+export { default as NavButton } from './NavButton.svelte';
 export { default as PageTitle } from './PageTitle.svelte';
 export { default as SectionHeader } from './SectionHeader.svelte';
 export { default as SiteCard } from './SiteCard.svelte';
